moved UiText to core.presentation.util
