+++ app/src/main/java/com/aarevalo/tasky/auth/domain/model/PasswordValidationResult.kt
@@ -1,6 +1,6 @@
 package com.aarevalo.tasky.auth.domain.model
 
-import com.aarevalo.tasky.core.util.UiText
+import com.aarevalo.tasky.core.presentation.util.UiText
 
 data class PasswordValidationResult(
     val isValid: Boolean,
