+++ app/src/main/java/com/aarevalo/tasky/core/presentation/util/UiEvent.kt
@@ -1,4 +1,4 @@
-package com.aarevalo.tasky.core.util
+package com.aarevalo.tasky.core.presentation.util
 
 sealed class UiEvent {
     data object Success: UiEvent()
